Fix divide-by-zero in MCTS fixed_log overflow
In fixed_log, the divisor (v + (1U << FIXED_SCALE_BITS)) is computed in
32-bit arithmetic. When v = 0xFFFF0000 (n_total=65535, so n_total<<16),
the addition overflows to zero, triggering a kernel Oops (divide error)
inside the kxod workqueue via ai_one_work_func -> play_agent_move ->
mcts.

This crashes the kworker, leaves the module refcount elevated, and makes
rmmod hang forever, which is CI timeout seen in integration-tests job.

Cast v to u64 before the addition so denominator is computed in 64-bit
and cannot wrap to zero.

+++ src/mcts.c
@@ -63,8 +63,8 @@ static fixed_point_t fixed_log(fixed_point_t v)
         numerator = (1U << 31) - numerator;
     }
 
-    fixed_point_t y =
-        ((u64) numerator << FIXED_SCALE_BITS) / (v + (1U << FIXED_SCALE_BITS));
+    fixed_point_t y = ((u64) numerator << FIXED_SCALE_BITS) /
+                      ((u64) v + (1U << FIXED_SCALE_BITS));
 
     fixed_point_t ans = 0U;
     for (unsigned i = 1; i < 20; i += 2) {
